refactor!: Make shared base package name be more specific (#3217)
Makes shared module package name to be more specific.
com.vaadin.flow.component package name already used in flow-server
so it produces a conflicts when used in OSGi environments.
Also, export directive was missing, which also produces errors.

Fixes vaadin/base-starter-flow-osgi#224
