+++ vaadin-flow-components-shared-parent/vaadin-flow-components-base/src/test/java/com/vaadin/flow/component/shared/SlotUtilsTest.java
@@ -13,8 +13,9 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.vaadin.flow.component;
+package com.vaadin.flow.component.shared;
 
+import com.vaadin.flow.component.HasElement;
 import com.vaadin.flow.dom.Element;
 import com.vaadin.flow.dom.ElementFactory;
